pyproject.toml
[project]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = [
    {name = "rodrigodsn",email = "rodrigo.barros.dsn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
pytest = "^8.4.0"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
python-dotenv = "^1.0.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview= true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = "ruff check ."
pre_format = "ruff check . --fix"
format = "ruff format"
run = 'uvicorn fast_zero.app:app --reload'
pre_test = 'task lint'
test = "pytest -s -x --cov=fast_zero -vv"
post_test = 'coverage html'
